:u6e80: fix(timestamp): fix time溢出问题
修改为记录时间ISO戳,字段为timestamp

BREAKING CHANGE: 时间戳字段和格式更改  record_time(number) ==> timestamp (string)
