+++ package.json
@@ -15,7 +15,8 @@
         "build": "npm run build:es5 && npm run build:es6",
         "build:es5": "tsc --target es5 -d --outDir dist",
         "build:es6": "tsc -m es6 --target es6 --outDir es",
-        "pub": "npm run build && npm publish --access public",
+        "prepub": "npm run build",
+        "pub": "npm publish --access public",
         "test": "npm run build"
     },
     "repository": {
