+++ src/log/index.ts
@@ -1,3 +1,3 @@
 export { ILogger, LogLevel, isLogLevel, AllLevels } from './ILogger';
 export { LogManager } from './log-manager';
-export { LogObject, LogTransformFunction,ReportAnalytics } from './report-analytics';
+export { LogObject, ReportAnalytics } from './report-analytics';
