+++ src/time/index.ts
@@ -1,2 +1,2 @@
 export { ITimer } from './ITimer';
-export { ReportAnalytics, PerformanceObject, PerformanceTransformFunction } from './report-analytics';
+export { ReportAnalytics, PerformanceObject } from './report-analytics';
